KPI popout component
Shows KPI as image with downloadable PDF and link to the D2D KPI in a new
tab/window.

The component has some variants depending on properties given to it. Not all CAP
goals have KPIs and not all KPI CAP goals have URLs to D2D (yet) so fallback to
PDF in that case.